moved cli
