+++ api/configuration/configuration.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 from typing import *
 from functools import cached_property
-from shared.singleton import Singleton
+from api.shared.singleton import Singleton
 from .variables import Variables
 from .environment import Environment
 from .environment.types import TEnvironment
