+++ api/shared/dependency/exceptions.py
