+++ api/shared/hash.py
