+++ api/shared/iterable.py
