+++ api/shared/singleton.py
